# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisJobPageResponse": {
        "description": "One paginated collection of analysis jobs.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisJobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "AnalysisJobPageResponse",
        "type": "object"
      },
      "AnalysisJobResponse": {
        "description": "An analysis job resource.",
        "examples": [
          {
            "analyzer": "caesar-bruteforce",
            "completed_at": "2026-09-30T12:00:00.004Z",
            "created_at": "2026-09-30T12:00:00Z",
            "id": "0b8f6a3e-2d43-4c1e-9f55-6a1d2b7c9e10",
            "language": "english",
            "result": {
              "analyzer": "caesar-bruteforce",
              "candidates": [
                {
                  "explanation": "Shift 3 scored -69.2637 against english.",
                  "key": "3",
                  "rank": 1,
                  "score": -69.26,
                  "text": "THE QUICK BROWN FOX JUMPS OVER THE LAZY DOG AND KEEPS RUNNING THROUGH THE FIELD"
                }
              ],
              "kind": "ranked-candidates",
              "language": "english",
              "language_version": "letter-frequency-v1"
            },
            "status": "succeeded",
            "updated_at": "2026-09-30T12:00:00.004Z",
            "version": 3
          }
        ],
        "properties": {
          "analyzer": {
            "title": "Analyzer",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "frequency-report": "#/components/schemas/FrequencyReportResult",
                    "ranked-candidates": "#/components/schemas/RankedCandidatesResult",
                    "vigenere-key-analysis": "#/components/schemas/VigenereKeyAnalysisResult"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RankedCandidatesResult"
                  },
                  {
                    "$ref": "#/components/schemas/FrequencyReportResult"
                  },
                  {
                    "$ref": "#/components/schemas/VigenereKeyAnalysisResult"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "analyzer",
          "language",
          "created_at",
          "updated_at",
          "completed_at",
          "error_code",
          "version",
          "result"
        ],
        "title": "AnalysisJobResponse",
        "type": "object"
      },
      "AnalysisRequest": {
        "additionalProperties": false,
        "description": "Ciphertext and analyzer selection for a keyless analysis.",
        "examples": [
          {
            "analyzer": "caesar-bruteforce",
            "language": "english",
            "text": "WKH TXLFN EURZQ IRA MXPSV RYHU WKH ODCB GRJ DQG NHHSV UXQQLQJ WKURXJK WKH ILHOG"
          }
        ],
        "properties": {
          "analyzer": {
            "description": "Analyzer to run. Supported: caesar-bruteforce, substitution-frequency, vigenere-frequency, homophonic-distribution.",
            "maxLength": 64,
            "minLength": 1,
            "title": "Analyzer",
            "type": "string"
          },
          "language": {
            "default": "english",
            "description": "Language model for scoring. Supported: english.",
            "maxLength": 64,
            "minLength": 1,
            "title": "Language",
            "type": "string"
          },
          "text": {
            "description": "Ciphertext to analyze.",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "analyzer",
          "text"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisStatus": {
        "description": "Lifecycle states of an analysis job.",
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
      "CandidateModel": {
        "description": "A ranked candidate key and plaintext.",
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Language score; null when the text has no scorable symbols.",
            "title": "Score"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "score",
          "key",
          "text",
          "explanation"
        ],
        "title": "CandidateModel",
        "type": "object"
      },
...
      "FrequencyReportResult": {
        "description": "Symbol or token frequency report.",
        "properties": {
          "analyzer": {
            "title": "Analyzer",
            "type": "string"
          },
          "comparisons": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Comparisons",
            "type": "object"
          },
          "kind": {
            "const": "frequency-report",
            "default": "frequency-report",
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "language_version": {
            "title": "Language Version",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "symbol_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Symbol Counts",
            "type": "object"
          },
          "symbol_frequencies": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Symbol Frequencies",
            "type": "object"
          }
        },
        "required": [
          "analyzer",
          "language",
          "language_version",
          "symbol_counts",
          "symbol_frequencies",
          "comparisons",
          "notes"
        ],
        "title": "FrequencyReportResult",
        "type": "object"
      },
      "KeyLengthCandidateModel": {
        "description": "A ranked Vigenere key-length estimate.",
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "key_length": {
            "title": "Key Length",
            "type": "integer"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "key_length",
          "score",
          "explanation"
        ],
        "title": "KeyLengthCandidateModel",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 problem details with machine-readable extensions.",
        "examples": [
          {
            "code": "invalid-key",
            "correlation_id": "4f1c2a8e-6c1b-4a51-9d7e-2b0f3f5b8c11",
            "detail": "Caesar key must be an integer.",
            "instance": "/v1/ciphers/caesar/encrypt",
            "status": 422,
            "title": "Invalid key",
            "type": "urn:codebreakers:problem:invalid-key"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "description": "Matches the X-Request-ID header.",
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "description": "Explanation specific to this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Present for schema validation failures.",
            "title": "Errors"
          },
          "instance": {
            "description": "Request path that produced the problem.",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlation_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "RankedCandidatesResult": {
        "description": "Candidates ranked by language score.",
        "properties": {
          "analyzer": {
            "title": "Analyzer",
            "type": "string"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateModel"
            },
            "title": "Candidates",
            "type": "array"
          },
          "kind": {
            "const": "ranked-candidates",
            "default": "ranked-candidates",
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "language_version": {
            "title": "Language Version",
            "type": "string"
          }
        },
        "required": [
          "analyzer",
          "language",
          "language_version",
          "candidates"
        ],
        "title": "RankedCandidatesResult",
        "type": "object"
      },
...
      "ValidationIssue": {
        "description": "A single schema validation failure, without the offending input value.",
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "VigenereKeyAnalysisResult": {
        "description": "Vigenere key-length estimates and candidate keys.",
        "properties": {
          "analyzer": {
            "title": "Analyzer",
            "type": "string"
          },
          "candidate_keys": {
            "items": {
              "$ref": "#/components/schemas/CandidateModel"
            },
            "title": "Candidate Keys",
            "type": "array"
          },
          "key_lengths": {
            "items": {
              "$ref": "#/components/schemas/KeyLengthCandidateModel"
            },
            "title": "Key Lengths",
            "type": "array"
          },
          "kind": {
            "const": "vigenere-key-analysis",
            "default": "vigenere-key-analysis",
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "language_version": {
            "title": "Language Version",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "minimum_useful_ciphertext_length": {
            "title": "Minimum Useful Ciphertext Length",
            "type": "integer"
          }
        },
        "required": [
          "analyzer",
          "language",
          "language_version",
          "key_lengths",
          "candidate_keys",
          "minimum_useful_ciphertext_length",
          "limitations"
        ],
        "title": "VigenereKeyAnalysisResult",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Encrypt, decrypt, and analyze text with classical ciphers. These ciphers are historically broken and provide no confidentiality; never use them to protect real secrets.",
    "title": "Codebreakers API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analyses": {
      "get": {
        "description": "Return retained jobs newest first using offset and limit pagination.\n\nJobs are not tied to a user, so a shared deployment disables listing and\nanswers `404`; each job is then reachable only through its own URL.",
        "operationId": "list_analyses",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List analyses",
        "tags": [
          "analyses"
        ]
      },
      "post": {
        "description": "Queue a keyless analysis and return the pending job resource.\n\nA worker runs the analysis asynchronously. Poll the `Location` URL until\n`status` is `succeeded`, `failed`, or `cancelled`. Clients over their\nsubmission rate get `429`, and a full job backlog answers `503`; both\ninclude `Retry-After`. The request's trace\ncontext (or a valid inbound W3C `traceparent` header when tracing is off)\nis propagated to the worker with the correlation ID.",
        "operationId": "create_analysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit an analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}": {
      "get": {
        "description": "Return a previously submitted analysis job and its current status.\n\nJobs are durable when PostgreSQL is configured; otherwise they are held in\nprocess memory and the oldest are evicted at capacity.",
        "operationId": "get_analysis",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get an analysis",
        "tags": [
          "analyses"
        ]
      }
    },
...
  },
...
}